LAB-2: added data class; tweet parsing is now in separate class; added grouping resulting tweets by hour

+++ SELab2/src/main/java/degtiarenko/hw/tweets/TweetStatisticService.java
@@ -3,5 +3,5 @@ package degtiarenko.hw.tweets;
 import java.util.List;
 
 public interface TweetStatisticService {
-    List<Integer> getHashTagPopularity(String hashTag, int hours) throws Exception;
+    List<Long> getHashTagPopularity(String hashTag, int hours) throws Exception;
 }
